Fix to the facebook reauth problem
After doing some research, I found stackoverflow posts that indicated
what we wanted was actually impossible.

https://stackoverflow.com/questions/16967356/why-my-app-still-visible-in-task-manager-after-finishing-it
"Pressing (and holding) HOME button doesn't open task manager. This is
some sort of "Recent Apps" section which shows you recently opened apps.
Actually it is a bit smarter than that - it can restore last visited
activity, but since you have only one activity, this shouldn't be really
important for you

So basically even if you can see your app in this list, it doesn't mean
it is actually running."

I figured out a way so that when the user clicked on the old activity in
the task list it would just return to desktop, but that wasn't very
good, and so I made it so that if the user actually clicks on that it
just takes them to the main screen.
I also removed a "finish()" call because it was preventing successful
reauthentication.